refactor(provider)!: node-side block heights are u64
get_block_number and Health.block_height carried Substrate's unsigned block number as i64, inherited from the indexer types where GraphQL Int forces signedness. Node-side heights now stay u64 end to end, matching the archive_v1 reads. Indexer-facing types (Block.height, the wallet's sync cursors) are unchanged: they mirror the GraphQL schema.

+++ crates/midnight-provider/tests/devnet.rs
@@ -98,7 +98,7 @@ async fn finalized_height_trails_the_best_head() {
     let best = p.get_block_number().await.unwrap();
     assert!(finalized > 0);
     assert!(
-        finalized as i64 <= best,
+        finalized <= best,
         "finalized {finalized} must not exceed best {best}"
     );
 }
